ex01 done

+++ C00/ex01/src/classes/Contact.hpp
@@ -18,7 +18,7 @@ class Contact
 	private:
 		std::string _getTruncString(std::string);
 		
-		int _columnSize;
+		size_t _columnSize;
 		std::string _firstName;
 		std::string _lastName;
 		std::string _nickName;
